requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==5.0.0
pluggy==0.8.1
py==1.7.0
pytest==4.2.0
python-dotenv==0.10.1
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
